# docs/esp/openapi.yaml
openapi: 3.0.0
info:
  title: ESP - News Pushing Pipeline API
  description: |
    Unified event query API over three data sources:
    - MongoDB (market news from Polygon.io)
    - SQLite (Benzinga earnings calendar)
    - SQLite (Nasdaq economic events calendar)

    ## Event Types
    - `macro_calendar` — Economic calendar events (FOMC, CPI, NFP, etc.)
    - `earnings` — Earnings announcements
    - `breaking_news` — Breaking market news (< 2h old)
    - `daily_news` — General market news

    ## Pagination
    POST endpoints use cursor-based pagination. Pass `next_cursor` from a response
    as `cursor` in the next request to fetch the next page.
  version: 0.1.0
  contact:
    name: QFinZero
servers:
  - url: http://localhost:19702
    description: Local development server
paths:
...
  /esp/events/query:
    post:
      operationId: queryEvents
      summary: Unified event query
      description: |
        Query events across all data sources with mode-based filtering.
        Modes: `upcoming` (future events), `just_happened` (recent events), `window` (time range).
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventQueryRequest'
      responses:
        '200':
          description: Paginated event results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /esp/events/{event_id}:
    get:
      operationId: getEvent
      summary: Get single event
      tags: [events]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
  /esp/events/stream:
    post:
      operationId: streamEvents
      summary: Incremental event polling
      description: Returns events since the cursor position for incremental polling.
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamRequest'
      responses:
        '200':
          description: Paginated event results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  # ── Triggers ──────────────────────────────────────────────────
...
  /esp/calendar/econ:
    post:
      operationId: econCalendar
      summary: Economic calendar
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EconCalendarRequest'
      responses:
        '200':
          description: Economic events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /esp/calendar/earnings:
    post:
      operationId: earningsCalendar
      summary: Earnings calendar
      tags: [calendar]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EarningsCalendarRequest'
      responses:
        '200':
          description: Earnings events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /esp/calendar/coverage:
    get:
      operationId: calendarCoverage
      summary: Calendar data coverage statistics
      description: |
        Returns coverage statistics for earnings and economic events data,
        including daily counts, missing dates, and breakdowns.
      tags: [calendar]
      parameters:
        - name: days
          in: query
          required: false
          description: Lookback window in days
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 30
      responses:
        '200':
          description: Coverage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoverageResponse'

  # ── News ──────────────────────────────────────────────────────
...
  /esp/news/stats:
    get:
      operationId: newsStats
      summary: News statistics
      description: |
        Returns news statistics including daily counts, top tickers,
        top publishers, and duplicate detection rates.
      tags: [news]
      parameters:
        - name: days
          in: query
          required: false
          description: Lookback window in days
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: News statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsStatsResponse'
        '503':
          description: MongoDB unavailable
  /esp/news/export:
    get:
      operationId: exportNews
      summary: Export news articles
      description: Export news articles as JSONL or CSV. Limited to 10,000 rows.
      tags: [export]
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [jsonl, csv]
            default: jsonl
        - name: tickers
          in: query
          required: false
          description: Comma-separated ticker filter
          schema:
            type: string
        - name: start
          in: query
          required: false
          description: Start datetime (ISO format)
          schema:
            type: string
        - name: end
          in: query
          required: false
          description: End datetime (ISO format)
          schema:
            type: string
      responses:
        '200':
          description: Exported data
          content:
            application/x-ndjson:
              schema:
                type: string
            text/csv:
              schema:
                type: string
        '400':
          description: Result set exceeds export limit
        '503':
          description: MongoDB unavailable

  # ── Export (Calendar) ─────────────────────────────────────────
  /esp/calendar/earnings/export:
    get:
      operationId: exportEarnings
      summary: Export earnings data
      description: Export earnings calendar data as JSONL or CSV. Limited to 10,000 rows.
      tags: [export]
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [jsonl, csv]
            default: csv
        - name: start
          in: query
          required: false
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: ticker
          in: query
          required: false
          description: Comma-separated ticker filter
          schema:
            type: string
      responses:
        '200':
          description: Exported data
        '400':
          description: Result set exceeds export limit
  /esp/calendar/economic/export:
    get:
      operationId: exportEconomic
      summary: Export economic events data
      description: Export economic calendar data as JSONL or CSV. Limited to 10,000 rows.
      tags: [export]
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [jsonl, csv]
            default: csv
        - name: start
          in: query
          required: false
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: false
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: country
          in: query
          required: false
          description: Country filter (default United States)
          schema:
            type: string
      responses:
        '200':
          description: Exported data
        '400':
          description: Result set exceeds export limit

  # ── Timeline ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        event_id:
          type: string
        event_type:
          type: string
          enum: [macro_calendar, earnings, breaking_news, daily_news]
        title:
          type: string
        time_utc:
          type: string
          format: date-time
        importance:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [scheduled, occurred, updated]
        tickers:
          type: array
          items:
            type: string
        country:
          type: string
          default: "US"
        snippet:
          type: string
        payload:
          type: object
        source:
          type: string
        source_id:
          type: string

    # ── Request Models ────────────────────────────────────────────
    EventQueryRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [upcoming, just_happened, window]
          default: upcoming
        start_utc:
          type: string
          format: date-time
        end_utc:
          type: string
          format: date-time
        horizon_minutes:
          type: integer
          default: 60
        event_types:
          type: array
          items:
            type: string
            enum: [macro_calendar, earnings, breaking_news, daily_news]
        tickers:
          type: array
          items:
            type: string
        min_importance:
          type: string
          enum: [low, medium, high]
        limit:
          type: integer
          default: 50
        cursor:
          type: string
        view:
          type: string
          enum: [compact, full]
          default: compact
        now_utc:
          type: string
          format: date-time
    StreamRequest:
      type: object
      properties:
        cursor:
          type: string
          description: "Cursor for pagination. Special value 'HEAD' means start from the latest position (may return empty if no new events buffered). Pass the 'next_cursor' from a previous response to continue from where you left off."
        event_types:
          type: array
          items:
            type: string
        tickers:
          type: array
          items:
            type: string
        limit:
          type: integer
          default: 50
        now_utc:
          type: string
...
    EconCalendarRequest:
      type: object
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        min_importance:
          type: string
          enum: [low, medium, high]
        limit:
          type: integer
          default: 100
        cursor:
          type: string
        now_utc:
          type: string
    EarningsCalendarRequest:
      type: object
      properties:
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        tickers:
          type: array
          items:
            type: string
        min_importance:
          type: integer
          minimum: 0
          maximum: 5
          default: 0
        limit:
          type: integer
          default: 100
        cursor:
          type: string
        now_utc:
          type: string
...
    PaginatedResponse:
      type: object
      properties:
        server_time_utc:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next_cursor:
          type: string
          nullable: true
...
    NewsStatsResponse:
      type: object
      properties:
        total_count:
          type: integer
        date_range:
          type: object
          properties:
            earliest:
              type: string
              nullable: true
            latest:
              type: string
              nullable: true
        daily_counts:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
        top_tickers:
          type: array
          items:
            type: object
            properties:
              ticker:
                type: string
              count:
                type: integer
        top_publishers:
          type: array
          items:
            type: object
            properties:
              publisher:
                type: string
              count:
                type: integer
        duplicate_stats:
          type: object
          properties:
            by_url:
              $ref: '#/components/schemas/DuplicateStats'
            by_title:
              $ref: '#/components/schemas/DuplicateStats'
    DuplicateStats:
      type: object
      properties:
        total:
          type: integer
        unique:
          type: integer
        duplicate_rate:
          type: number
          format: double
    CoverageResponse:
      type: object
      properties:
        earnings:
          type: object
          properties:
            date_range:
              type: object
              properties:
                start:
                  type: string
                end:
                  type: string
            total_records:
              type: integer
            daily_counts:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                  count:
                    type: integer
            missing_dates:
              type: array
              items:
                type: string
            by_importance:
              type: object
              additionalProperties:
                type: integer
        econ_events:
          type: object
          properties:
            date_range:
              type: object
              properties:
                start:
                  type: string
                end:
                  type: string
            total_records:
              type: integer
            daily_counts:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                  count:
                    type: integer
            missing_dates:
              type: array
              items:
                type: string
            by_country:
              type: object
              additionalProperties:
                type: integer
            by_type_top10:
              type: array
              items:
                type: object
                properties:
                  event_type:
                    type: string
                  count:
                    type: integer
...